+++ ValheimEnforcer/ValheimEnforcer.cs
@@ -24,7 +24,7 @@ namespace ValheimEnforcer
     {
         public const string PluginGUID = "MidnightsFX.ValheimEnforcer";
         public const string PluginName = "ValheimEnforcer";
-        public const string PluginVersion = "0.9.1";
+        public const string PluginVersion = "0.10.0";
 
         internal static ManualLogSource Log;
         internal ValConfig cfg;